+++ charts/core/src/index.ts
@@ -1,5 +1,4 @@
 export { Chart, type ChartProps, ChartStates } from './Chart';
-export { ChartCard, type ChartCardProps, ChartCardStates } from './ChartCard';
 export {
   EventMarkerLine,
   type EventMarkerLineProps,