+++ charts/legend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lg-charts/legend
 
+## 0.2.0
+
+### Minor Changes
+
+- 206278b0f: Adds `Legend` component
+
+### Patch Changes
+
+- Updated dependencies [206278b0f]
+- Updated dependencies [206278b0f]
+- Updated dependencies [206278b0f]
+  - @lg-charts/colors@0.2.2
+  - @leafygreen-ui/checkbox@15.2.0
+  - @lg-charts/series-provider@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes